requirements.txt
python-decouple>=3.3
Flask>=1.1.2
Flask-RESTful>=0.3.8
flask-mongoengine>=0.9.5
Flask-Bcrypt>=0.7.1
Flask-JWT-Extended>=3.24.1